Rename UserInterface and UserRepository
Rename UserInterface to UserRepository and
UserRepository to UserInMemoryRepository
